General clean-up of old project
Whats new:
  - Now `Damm32.c` is `Damm.cpp`!
  - Major clean-up of old MarkDown syntax.
  - Build for x86-64 and arm-gnueabi.

More changes soon!
